Split StellarPaymentObserver
